requirements.txt
einops
graphviz
hydra-core
loguru
numpy
opencv-python
Pillow
pycocotools
pytest
pyyaml
requests
rich
torch
torchvision
tqdm
wandb
